+++ ModHelperData.cs
@@ -3,7 +3,7 @@
 internal static class ModHelperData
 {
     public const string WorksOnVersion = "55.2";
-    public const string Version = "4.4.5";
+    public const string Version = "4.4.6";
     public const string Name = "Ability Choice";
 
     public const string Description =